Expand input path globs (*, **)
This feature is mainly for Windows users. On *nix systems unquoted
input paths will already be expanded by most shells, but this is not
a feature on Windows, so we perform the expansion for them.

Unfortunately the Go-stdlib doesn't support double-star (**) format,
which is common in most shells.

https://github.com/golang/go/issues/11862

So the following library was a good fit (it also allows pattern matching like
*.{jpg,png} which some other Go globbing libraries don't support).

https://github.com/bmatcuk/doublestar