Fix skipping existing images
Unless reprocessing is allowed, existing images should be skipped
to preserve credits.

This functionality was broken when the auto-upgrade from PNG to ZIP
was introduced. It ended up checking if the .zip file already exists,
which it wouldn't because a .png is emitted when the .zip format is used.

Fixes https://github.com/remove-bg/go/issues/16

Source of bug:
https://github.com/remove-bg/go/pull/7/commits/e40648c6b73ec916a90eaf253c8e152b3560312f